refactor: rename project to rich_trader
