+++ manage.py
@@ -9,7 +9,7 @@ import dotenv
 def main():
     """Run administrative tasks."""
     dotenv.read_dotenv()
-    os.environ.setdefault("DJANGO_SETTINGS_MODULE", "rich_kraken.settings.local")
+    os.environ.setdefault("DJANGO_SETTINGS_MODULE", "rich_trader.settings.local")
     try:
         from django.core.management import execute_from_command_line
     except ImportError as exc:
