+++ requirements.in
@@ -11,8 +11,6 @@ git+https://github.com/sebatyler/django-s3-sqlite.git@main#egg=django-s3-sqlite
 
 jinja2==3.1.3
 
-pykrakenapi
-
 python-telegram-bot==13.13
 
 zappa==0.58.0
