+++ rich/jinja2/index.html
@@ -1,5 +1,5 @@
 <head>
-  <title>Kraken</title>
+  <title>Rich Trader</title>
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <script src="https://cdn.tailwindcss.com"></script>
   <script src="https://cdn.jsdelivr.net/npm/chart.js"></script>
