+++ rich/tests.py
