+++ rich_trader/asgi.py
@@ -1,5 +1,5 @@
 """
-ASGI config for rich_kraken project.
+ASGI config for rich_trader project.
 
 It exposes the ASGI callable as a module-level variable named ``application``.
 
@@ -10,10 +10,11 @@ https://docs.djangoproject.com/en/4.1/howto/deployment/asgi/
 import os
 
 import dotenv
+
 from django.core.asgi import get_asgi_application
 
 dotenv.read_dotenv(os.path.join(os.path.dirname(os.path.dirname(__file__)), ".env"))
 
-os.environ.setdefault("DJANGO_SETTINGS_MODULE", "rich_kraken.settings.prod")
+os.environ.setdefault("DJANGO_SETTINGS_MODULE", "rich_trader.settings.prod")
 
 application = get_asgi_application()
