+++ rich_trader/settings/base.py
@@ -1,5 +1,5 @@
 """
-Django settings for rich_kraken project.
+Django settings for rich_trader project.
 
 Generated by 'django-admin startproject' using Django 4.1.
 
@@ -30,7 +30,7 @@ SECRET_KEY = "django-insecure-)fo36l*+#9#ng=@a&g!u*^c^42qgz(ka1%8$m73hp++91y+x#*
 # SECURITY WARNING: don't run with debug turned on in production!
 DEBUG = True
 
-DJANGO_SETTINGS_MODULE = os.getenv("DJANGO_SETTINGS_MODULE", "rich_kraken.settings.local")
+DJANGO_SETTINGS_MODULE = os.getenv("DJANGO_SETTINGS_MODULE", "rich_trader.settings.local")
 PROJECT_NAME, *_, ENV = DJANGO_SETTINGS_MODULE.split(".")
 
 ALLOWED_HOSTS = ["*"]
@@ -47,7 +47,7 @@ INSTALLED_APPS = [
     "django.contrib.messages",
     "django.contrib.staticfiles",
     "django_extensions",
-    "kraken",
+    "rich",
 ]
 
 MIDDLEWARE = [
@@ -60,7 +60,7 @@ MIDDLEWARE = [
     "django.middleware.clickjacking.XFrameOptionsMiddleware",
 ]
 
-ROOT_URLCONF = "rich_kraken.urls"
+ROOT_URLCONF = "rich_trader.urls"
 
 TEMPLATES = [
     {
@@ -68,7 +68,7 @@ TEMPLATES = [
         "DIRS": [],
         "APP_DIRS": True,
         "OPTIONS": {
-            "environment": "rich_kraken.jinja2.environment",
+            "environment": "rich_trader.jinja2.environment",
         },
     },
     {
@@ -86,7 +86,7 @@ TEMPLATES = [
     },
 ]
 
-WSGI_APPLICATION = "rich_kraken.wsgi.application"
+WSGI_APPLICATION = "rich_trader.wsgi.application"
 
 
 # Database
@@ -95,7 +95,7 @@ if os.getenv("USE_S3_SQLITE", "0") == "1":
     DATABASES = {
         "default": {
             "ENGINE": "django_s3_sqlite",
-            "NAME": "rich_kraken.db",
+            "NAME": "rich_trader.db",
             "BUCKET": "sebatyler-dev",
         },
     }
