+++ rich_trader/settings/local.py
