+++ rich_trader/settings/test.py
