+++ rich_trader/urls.py
@@ -1,4 +1,4 @@
-"""rich_kraken URL Configuration
+"""rich_trader URL Configuration
 
 The `urlpatterns` list routes URLs to views. For more information please see:
     https://docs.djangoproject.com/en/4.1/topics/http/urls/
@@ -17,7 +17,7 @@ Including another URLconf
 from django.contrib import admin
 from django.urls import path
 
-from kraken.views import IndexView
+from rich.views import IndexView
 
 urlpatterns = [
     path("_a/", admin.site.urls),
