+++ rich_trader/wsgi.py
@@ -1,5 +1,5 @@
 """
-WSGI config for rich_kraken project.
+WSGI config for rich_trader project.
 
 It exposes the WSGI callable as a module-level variable named ``application``.
 
@@ -10,10 +10,11 @@ https://docs.djangoproject.com/en/4.1/howto/deployment/wsgi/
 import os
 
 import dotenv
+
 from django.core.wsgi import get_wsgi_application
 
 dotenv.read_dotenv(os.path.join(os.path.dirname(os.path.dirname(__file__)), ".env"))
 
-os.environ.setdefault("DJANGO_SETTINGS_MODULE", "rich_kraken.settings.prod")
+os.environ.setdefault("DJANGO_SETTINGS_MODULE", "rich_trader.settings.prod")
 
 application = get_wsgi_application()
