+++ upload.sh
@@ -1,3 +1,3 @@
 #!/bin/bash
 
-aws --profile sebatyler s3 cp .env.prod s3://sebatyler-dev/rich_kraken/.env.prod
+aws --profile sebatyler s3 cp .env.prod s3://sebatyler-dev/rich_trader/.env.prod
